Sql sentences needs revision
