--- UC1/src/main/java/com/uclm/esi/isoft2/a04/Persistence/BookingDao.java
@@ -1,48 +1,0 @@
-package Persistence;
-
-import Domain.*;
-
-public class BookingDao {
-
-	public Booking[] readAllBookings() {
-		// TODO - implement BookingDao.readAllBookings
-		throw new UnsupportedOperationException();
-	}
-
-	/**
-	 * 
-	 * @param booking
-	 */
-	public void readBooking(Booking booking) {
-		// TODO - implement BookingDao.readBooking
-		throw new UnsupportedOperationException();
-	}
-
-	/**
-	 * 
-	 * @param booking
-	 */
-	public int createBooking(Booking booking) {
-		// TODO - implement BookingDao.createBooking
-		throw new UnsupportedOperationException();
-	}
-
-	/**
-	 * 
-	 * @param booking
-	 */
-	public int updateBooking(Booking booking) {
-		// TODO - implement BookingDao.updateBooking
-		throw new UnsupportedOperationException();
-	}
-
-	/**
-	 * 
-	 * @param booking
-	 */
-	public int deleteBooking(Booking booking) {
-		// TODO - implement BookingDao.deleteBooking
-		throw new UnsupportedOperationException();
-	}
-
-}
