+++ UC1/src/main/java/com/uclm/esi/isoft2/a04/Persistence/TableDAO.java
@@ -1,6 +1,12 @@
-package Persistence;
+package com.uclm.esi.isoft2.a04.Persistence;
 
-import Domain.*;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.ResultSet;
+import java.sql.Statement;
+
+import com.uclm.esi.isoft2.a04.Domain.*;
+
 
 public class TableDAO {
 
@@ -14,8 +20,20 @@ public class TableDAO {
 	 * @param table
 	 */
 	public void readTable(TableImplementation table) {
-		// TODO - implement TableDAO.readTable
-		throw new UnsupportedOperationException();
+		
+		try {
+			String url = "jdbc:mysql://172.20.48.70:3306/XXXdbservice?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC";
+			Connection conn = DriverManager.getConnection(url,"","");
+			Statement statement = conn.createStatement();
+			ResultSet resultQuery;
+			
+			resultQuery = statement.executeQuery("SELECT  FROM Booking WHERE ");
+			
+		}
+		catch (Exception e) {
+			System.err.println("An exception has occur");
+			System.out.println(e.getMessage());
+		}
 	}
 
 	/**
@@ -23,9 +41,26 @@ public class TableDAO {
 	 * @param table
 	 */
 	public int createTable(TableImplementation table) {
-		// TODO - implement TableDAO.createTable
-		throw new UnsupportedOperationException();
+		
+		try {
+			String url = "jdbc:mysql://172.20.48.70:3306/XXXdbservice?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC";
+			Connection conn = DriverManager.getConnection(url,"","");
+			Statement statement = conn.createStatement();
+			ResultSet resultQuery;
+			
+			resultQuery = statement.executeQuery("DELETE FROM  WHERE id = ");
+			
+			return resultQuery.getType();
+			
+		}
+		catch (Exception e) {
+			System.err.println("An exception has occur");
+			System.out.println(e.getMessage());
+		}
+		return -1;
+		
 	}
+		
 
 	/**
 	 *
